+++ stwcs/tests/test_hstwcs.py
@@ -1,5 +1,5 @@
 from astropy.io import fits
-from ..wcsutil import hstwcs
+from stwcs.wcsutil import hstwcs, HSTWCS
 
 
 def test_radesys():
@@ -12,3 +12,13 @@ def test_radesys():
     assert hstwcs.determine_refframe(phdr) is None
     phdr['refframe'] = ' '
     assert hstwcs.determine_refframe(phdr) is None
+
+
+def test_prihdu_with_extver_no_extname():
+    hdulist = fits.HDUList([
+        fits.PrimaryHDU(header=fits.Header([('extver', 7)])),
+        fits.ImageHDU(header=fits.Header([('time', 6)]))
+    ])
+    extname = HSTWCS(hdulist).extname
+    assert extname == ('PRIMARY', 7)
+    assert hdulist[extname] is hdulist[0]